// apps/web/src/types/sanity.types.ts
export type Navigation = {
  _id: string;
  _type: "navigation";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  menus?: Array<{
    label: string;
    menu?: Array<{
      icon?: string;
      link?: {
        type?: "internalLink" | "link";
        label?: string;
        link?: {
          href?: string;
          alt?: string;
        };
        internalLink?: {
          reference?: {
            _ref: string;
            _type: "reference";
            _weak?: boolean;
            [internalGroqTypeReferenceTo]?: "page";
          } | {
            _ref: string;
            _type: "reference";
            _weak?: boolean;
            [internalGroqTypeReferenceTo]?: "legal";
          } | {
            _ref: string;
            _type: "reference";
            _weak?: boolean;
            [internalGroqTypeReferenceTo]?: "blog";
          };
          blank?: boolean;
          alt?: string;
        };
      };
      badge?: string;
      description?: string;
      _type: "navItem";
      _key: string;
    }>;
    _type: "menu";
    _key: string;
  } | {
    label: string;
    column?: Array<{
      label: string;
      menu?: Array<{
        icon?: string;
        link?: {
          type?: "internalLink" | "link";
          label?: string;
          link?: {
            href?: string;
            alt?: string;
          };
          internalLink?: {
            reference?: {
              _ref: string;
              _type: "reference";
              _weak?: boolean;
              [internalGroqTypeReferenceTo]?: "page";
            } | {
              _ref: string;
              _type: "reference";
              _weak?: boolean;
              [internalGroqTypeReferenceTo]?: "legal";
            } | {
              _ref: string;
              _type: "reference";
              _weak?: boolean;
              [internalGroqTypeReferenceTo]?: "blog";
            };
            blank?: boolean;
            alt?: string;
          };
        };
        badge?: string;
        description?: string;
        _type: "navItem";
        _key: string;
      }>;
      _type: "menu";
      _key: string;
    }>;
    _type: "megaMenu";
    _key: string;
  } | {
    type?: "internalLink" | "link";
    label?: string;
    link?: {
      href?: string;
      alt?: string;
    };
    internalLink?: {
      reference?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "page";
      } | {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "legal";
      } | {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "blog";
      };
      blank?: boolean;
      alt?: string;
    };
    _type: "link";
    _key: string;
  }>;
  buttons?: Array<{
    link?: {
      type?: "internalLink" | "link";
      label?: string;
      link?: {
        href?: string;
        alt?: string;
      };
      internalLink?: {
        reference?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "page";
        } | {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "legal";
        } | {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "blog";
        };
        blank?: boolean;
        alt?: string;
      };
    };
    hierarchy?: "primary" | "outline" | "link";
    icon?: string;
    iconPosition?: "start" | "end";
    _type: "button";
    _key: string;
  }>;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;